Log in to docker hub before pulling images
Docker Hub is implementing rate limiting for anonymous pulls. We need to
log in before we pull. That happens in two places

1) In all job to fetch the image that the CircleCI jobs run in. This is what
the new `auth` block handles.

2) In the build job to pull the base image defined in the application's dockerfile. This
is that the new "login to dockerhub" step handles,

We also still have to log in in the deploy job in order to push to
docker hub. There is no shared state between jobs,

h/t to jwhitlock who figured most of this out in https://github.com/mozilla/ichnaea/compare/6655b9b...f8da96dc28